% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature.R
\name{get_feature}
\alias{get_feature}
\title{Produce an SF object}
\usage{
get_feature(
  base.url = "https://kartta.hsy.fi/geoserver/wfs",
  typename = "asuminen_ja_maankaytto:Vaestotietoruudukko_2015"
)
}
\arguments{
\item{base.url}{WFS url, for example "https://kartta.hsy.fi/geoserver/wfs"}

\item{typename}{accepts feature names, e.g. "asuminen_ja_maankaytto:1000m_verkostobufferi"
No short form titles here, e.g. "1000m_verkostobufferi"!}
}
\value{
sf object
}
\description{
Produces an sf object for easy visualization
}
\details{
See list_features() for a list of all available features
}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
