% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_list.R
\name{get_hri_dataset_list}
\alias{get_hri_dataset_list}
\title{List HRI datasets}
\usage{
get_hri_dataset_list()
}
\description{
List HRI datasets
}
\details{
This function lists all available HRI datasets.
}
\keyword{internal}
