% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_list.R
\name{get_hri_dataset_metadata}
\alias{get_hri_dataset_metadata}
\title{Get HRI dataset metadata}
\usage{
get_hri_dataset_metadata(id)
}
\description{
Get HRI dataset metadata
}
\details{
This function fetches metadata for HRI datasets.
}
\keyword{internal}
