% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_stats.R
\name{get_hri_stats}
\alias{get_hri_stats}
\title{Helsinki Region Infoshare statistics API}
\usage{
get_hri_stats(query = "", verbose = TRUE)
}
\arguments{
\item{query}{A string, specifying the dataset to query}

\item{verbose}{logical. Should R report extra information on progress?
Default is TRUE}
}
\value{
multi-dimensional array
}
\description{
Retrieves data from the Helsinki Region Infoshare (HRI)
statistics API: \url{http://dev.hel.fi/stats/}.
Currently provides access to the \emph{aluesarjat} data:
\url{https://stat.hel.fi/pxweb/fi/Aluesarjat/}.
}
\details{
Current implementation is very simple.
You can either get the list of resources with query="",
or query for a specific resources and retrieve it in a
three-dimensional array form.
}
\examples{
\dontrun{
stats_array <- get_hri_stats("aluesarjat_a03s_hki_vakiluku_aidinkieli")
}

}
\references{
See citation("helsinki")
}
\seealso{
See \url{https://dev.hel.fi/apis/statistics/}{dev.hel.fi website} and
\url{http://dev.hel.fi/stats/}{API documentation} (in Finnish)
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}
