% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSY.R
\name{get_hsy-deprecated}
\alias{get_hsy-deprecated}
\title{Retrieve data from Helsinki Region Environmental Services.}
\usage{
get_hsy(which.data, which.year, data.dir, verbose)
}
\arguments{
\item{which.data}{A string. Specify the name of the retrieved HSY data set. Options: Vaestotietoruudukko; Rakennustietoruudukko; SeutuRAMAVA_kosa; SeutuRAMAVA_tila. These are documented in HSY data description document (see above).}

\item{which.year}{An integer. Specify the year for the data to be retrieved.}

\item{data.dir}{A string. Specify a temporary folder for storing downloaded data.}

\item{verbose}{logical. Should R report extra information on progress?}
}
\value{
Shape object (from SpatialPolygonsDataFrame class)
}
\description{
Retrieves data from Helsinki Region Environmental Services Authority.
}
\seealso{
\code{\link{helsinki-deprecated}}
}
\keyword{internal}
