% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_features.R
\name{get_interactive}
\alias{get_interactive}
\title{Interactively browse and select features}
\usage{
get_interactive()
}
\value{
feature Title (character)
}
\description{
Select wanted feature for use in other functions
}
\examples{
\dontrun{
selection <- interactive_list()
feature <- simple_get(type_name = selected)
ggplot(feature) +
  geom_sf()
}

}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
