% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSY.R
\name{get_vaestotietoruudukko}
\alias{get_vaestotietoruudukko}
\title{Produce an SF object: Vaestotietoruudukko}
\usage{
get_vaestotietoruudukko(year = NULL)
}
\arguments{
\item{year}{year as numeric from range 2015:2019}
}
\value{
sf object
}
\description{
Produces an sf object for Väestötietoruudukko (population grid)
}
\details{
Additional data not available here can be manually downloaded from HRI website: https://hri.fi/data/fi/dataset/vaestotietoruudukko

Years 2015-2019 are tested to work at the time of development. 
Additional years may be added in the future and older datasets may be removed
from the API. 

The current datasets can be listed with get_feature_list()
or select_feature().
}
\examples{
\dontrun{
pop_grid <- get_vaestotietoruudukko(year = 2015)
}

}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
