\docType{data}
\name{HRI.aluejakokartat}
\alias{HRI.aluejakokartat}
\title{HRI.aluejakokartat data documentation}
\format{list of SpatialPolygonsDataFrames}
\description{
  This data set contains information about area boundaries
  for Helsinki Region Helsinki Region maps from Helsinki
  Region Infoshare Contents: pks.pienalue: Detailed area
  boundary data
}
\author{
  Juuso Parkkinen \email{louhos@googlegroups.com}
}
\references{
  See cite(sorvi)
}
\keyword{data}
\keyword{datasets}
\keyword{misc}

