% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSY.R
\name{simple_get}
\alias{simple_get}
\title{Produce an SF object}
\usage{
simple_get(typeName = "asuminen_ja_maankaytto:Vaestotietoruudukko_2015")
}
\arguments{
\item{typeName}{accepts feature names, e.g. "asuminen_ja_maankaytto:1000m_verkostobufferi"
No short form titles here, e.g. "1000m_verkostobufferi"!}
}
\value{
sf object
}
\description{
Produces an sf object for easy visualization
}
\details{
See list_features() for a list of all available features
}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
