% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sf.R
\name{to_sf}
\alias{to_sf}
\title{Transform to sf-object}
\usage{
to_sf(api_obj)
}
\arguments{
\item{api_obj}{wfs api object}
}
\value{
sf object
}
\description{
Transform a wfs_api object into a sf object.
}
\details{
FMI API response object's XML (GML) content is temporarily wrtitten on disk
and then immediately read back in into a sf object.
}
\note{
For internal use, not exported.
}
\author{
Joona Lehtomäki <joona.lehtomaki@iki.fi>
}
\keyword{utilities}
