% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_map.R
\name{get_city_map}
\alias{get_city_map}
\title{Get city administrative regions}
\usage{
get_city_map(city = NULL, level = NULL)
}
\arguments{
\item{city}{The desired city. Options: Helsinki, Espoo, Vantaa, Kauniainen}

\item{level}{The desired administrative level. Options are 1, 2 and 3,
where 1 is the lowest level (= largest areas) and 3 is the highest level 
(= smallest areas)}
}
\value{
sf object
}
\description{
Sf object of city districts
}
\details{
See list_features() for a list of all available features
}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
