% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature.R
\name{get_feature}
\alias{get_feature}
\title{Produce an SF object}
\usage{
get_feature(
  base.url = "https://kartta.hsy.fi/geoserver/wfs",
  typename = "asuminen_ja_maankaytto:Vaestotietoruudukko_2015",
  CRS = 3879
)
}
\arguments{
\item{base.url}{WFS url, for example "https://kartta.hsy.fi/geoserver/wfs"}

\item{typename}{accepts feature names, e.g. "asuminen_ja_maankaytto:1000m_verkostobufferi"
No short form titles here, e.g. "1000m_verkostobufferi"!}

\item{CRS}{Default CRS is 3879 (or EPSG:3879), see ?sf::st_crs for other input options}
}
\value{
sf object
}
\description{
Produces an sf object for easy visualization
}
\seealso{
Use \code{\link{get_feature_list}} to list all available features
for a given WFS, \code{\link{select_feature}} for listing and selecting a
feature
}
\author{
Pyry Kantanen \href{mailto:pyry.kantanen@gmail.com}{pyry.kantanen@gmail.com}
}
