% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_features.R
\name{get_feature_list}
\alias{get_feature_list}
\title{Print all available Features}
\usage{
get_feature_list()
}
\value{
data frame
}
\description{
Basically "request=GetCapabilities" as a neat data frame.
}
\details{
Lists all <FeatureType> nodes.
}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
