% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/services.R
\name{get_linkedevents}
\alias{get_linkedevents}
\title{Access Helsinki Linked Events API}
\usage{
get_linkedevents(query, ...)
}
\arguments{
\item{query}{The API query as a string, one of "category", "event", "language", or "place".
For details, see http://api.hel.fi/linkedevents/v0.1/.}

\item{...}{Additional parameters to the API (optional).
For details, see http://api.hel.fi/linkedevents/v0.1/.}
}
\value{
List of results
}
\description{
Access the new Helsinki Linked Events API: http://api.hel.fi/linkedevents/v0.1/.
The API contains data from the Helsinki City Tourist & Convention Bureau,
the City of Helsinki Cultural Office and the Helmet metropolitan area public libraries.
For more API documentation and license information see the API link.
}
\examples{
events <- get_linkedevents(query="event")
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}

