% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSY.R
\name{get_rakennustietoruudukko}
\alias{get_rakennustietoruudukko}
\title{Produce an SF object: Rakennustietoruudukko}
\usage{
get_rakennustietoruudukko(year = NULL)
}
\arguments{
\item{year}{year as numeric from range 2015:2019}
}
\value{
sf object
}
\description{
Produces an sf object for Rakennustietoruudukko (building information grid)
}
\details{
Additional data not available here can be manually downloaded from HRI website: https://hri.fi/data/fi/dataset/rakennustietoruudukko
}
\examples{
\dontrun{
pop_grid <- get_rakennustietoruudukko(year = 2016)
}

}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
