% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_servicemap}
\alias{get_servicemap}
\title{Access Helsinki region Service Map API}
\usage{
get_servicemap(query, ...)
}
\arguments{
\item{query}{The API query as a string, for example "search", "service", or "unit".
For full list of available options and details, see http://api.hel.fi/servicemap/v1/.}

\item{...}{Additional parameters to the API (optional).
For details, see http://api.hel.fi/servicemap/v1/.}
}
\value{
List of results
}
\description{
Access the new Helsinki region Service Map (Paakaupunkiseudun Palvelukartta)
http://dev.hel.fi/servicemap/ data through the API: http://api.hel.fi/servicemap/v1/.
For more API documentation and license information see the API link.
}
\examples{
\donttest{ search.puisto <- get_servicemap(category="search", q="puisto") }
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}

