% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_servicemap}
\alias{get_servicemap}
\title{Access Helsinki region Service Map API}
\usage{
get_servicemap(category, ...)
}
\arguments{
\item{category}{API query category, for example "search", "service", or "unit".
For list of available options, see http://api.hel.fi/servicemap/v1/.}

\item{...}{Additional parameters to the API (optional). See details from the API documentation http://api.hel.fi/servicemap/v1/.}
}
\value{
List of results
}
\description{
Access the new Helsinki region Service Map (Paakaupunkiseudun Palvelukartta)
http://dev.hel.fi/servicemap/ data through the API: http://api.hel.fi/servicemap/v1/.
For more details and API documentation see http://api.hel.fi/servicemap/v1/.
}
\examples{
\dontrun{ pk.services <- get_servicemap("service") }
}
\author{
Juuso Parkkinen \email{louhos@googlegroups.com}
}

