% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{helsinki-package}
\alias{helsinki}
\alias{helsinki-package}
\title{Helsinki open data R tools}
\description{
Brief summary of the helsinki package
}
\details{
\tabular{ll}{
Package: \tab helsinki\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014\cr
License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr
}

Tools for accessing various open data sources in the Helsinki region in Finland.
}
\examples{
# library(helsinki)
}
\author{
Juuso Parkkinen, Leo Lahti and Joona Lehtomaki \email{louhos@googlegroups.com}
}
\references{
See citation("helsinki")
\url{https://github.com/rOpenGov/helsinki}
}
\keyword{package}

