\docType{package}
\name{helsinki-package}
\alias{helsinki}
\alias{helsinki-package}
\title{R Tools for helsinki API}
\description{
  Brief summary of the helsinki package
}
\details{
  \tabular{ll}{ Package: \tab helsinki\cr Type: \tab
  Package\cr Version: \tab See sessionInfo() or DESCRIPTION
  file\cr Date: \tab 2013\cr License: \tab FreeBSD\cr
  LazyLoad: \tab yes\cr }

  R Tools for helsinki API
}
\examples{
# library(helsinki)
}
\author{
  Leo Lahti, Juuso Parkkinen and Joona Lehtomaki
  \email{louhos@googlegroups.com}
}
\references{
  See citation("helsinki")
  \url{http://ropengov.github.com/helsinki}
}
\keyword{package}

