% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_list.R
\name{select_feature}
\alias{select_feature}
\title{Interactively browse and select features}
\usage{
select_feature()
}
\value{
feature Title (character)
}
\description{
Select wanted feature for use in other functions
}
\examples{
\dontrun{
selection <- select_feature()
feature <- get_feature(type_name = selected)
ggplot(feature) +
  geom_sf()
}

}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
